requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.10
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-heroku==0.3.1
django-jazzmin==2.5.0
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
threadpoolctl==3.1.0
urllib3==1.26.9
whitenoise==6.0.0
django-import-export==2.8.0

